+++ Userland/Libraries/LibWeb/CSS/Clip.cpp
@@ -5,7 +5,6 @@
  */
 
 #include "Clip.h"
-#include <LibWeb/CSS/StyleValue.h>
 
 namespace Web::CSS {
