+++ Userland/Libraries/LibWeb/CSS/StyleValue.cpp
@@ -777,31 +777,6 @@ Optional<CalculatedStyleValue::ResolvedType> CalculatedStyleValue::CalcSum::reso
     return resolve_sum_type(type, zero_or_more_additional_calc_products);
 }
 
-// https://www.w3.org/TR/CSS2/visufx.html#value-def-shape
-Gfx::FloatRect EdgeRect::resolved(Layout::Node const& layout_node, Gfx::FloatRect border_box) const
-{
-    // In CSS 2.1, the only valid <shape> value is: rect(<top>, <right>, <bottom>, <left>) where
-    // <top> and <bottom> specify offsets from the top border edge of the box, and <right>, and
-    // <left> specify offsets from the left border edge of the box.
-
-    // The value 'auto' means that a given edge of the clipping region will be the same as the edge
-    // of the element's generated border box (i.e., 'auto' means the same as '0' for <top> and
-    // <left>, the same as the used value of the height plus the sum of vertical padding and border
-    // widths for <bottom>, and the same as the used value of the width plus the sum of the
-    // horizontal padding and border widths for <right>, such that four 'auto' values result in the
-    // clipping region being the same as the element's border box).
-    auto left = border_box.left() + (left_edge.is_auto() ? 0 : left_edge.to_px(layout_node)).value();
-    auto top = border_box.top() + (top_edge.is_auto() ? 0 : top_edge.to_px(layout_node)).value();
-    auto right = border_box.left() + (right_edge.is_auto() ? border_box.width() : right_edge.to_px(layout_node)).value();
-    auto bottom = border_box.top() + (bottom_edge.is_auto() ? border_box.height() : bottom_edge.to_px(layout_node)).value();
-    return Gfx::FloatRect {
-        left,
-        top,
-        right - left,
-        bottom - top
-    };
-}
-
 Optional<CalculatedStyleValue::ResolvedType> CalculatedStyleValue::CalcNumberSum::resolved_type() const
 {
     auto maybe_type = first_calc_number_product->resolved_type();
